requirements.txt
langchain
langchain-community
llama-parse 
fastembed
qdrant_client
python-dotenv
langchain-groq
chainlit
fastembed
unstructured[md]
nest_asyncio
openai
langchain-openai